server/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.13.0
python-dotenv==1.1.0
Werkzeug==3.1.3
gunicorn
pydicom==2.3.1
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
opencv-python==4.10.0.84

